patch for platform __init__ to enable local execution of the GUI

+++ dream/platform/__init__.py
@@ -25,7 +25,6 @@ import urllib
 import xlrd
 import traceback
 import multiprocessing
-import Queue
 from dream.KnowledgeExtraction.DistributionFitting import DistFittest
 from dream.KnowledgeExtraction.ImportExceldata import Import_Excel
 
@@ -100,37 +99,35 @@ def positionGraph():
 class TimeoutError(Exception):
   pass
 
+#def runWithTimeout(func, timeout, *args, **kw):
 def runWithTimeout(func, timeout, *args, **kw):
+  return func(*args, **kw)
   queue = multiprocessing.Queue()
   process = multiprocessing.Process(
     target=_runWithTimeout,
     args=(queue, func, args, kw))
   process.start()
-  try:
-    return queue.get(timeout=timeout)
-  except Queue.Empty:
+  process.join(timeout)
+  if process.is_alive():
+    # process still alive after timeout, terminate it
     process.terminate()
+    process.join() 
+    import traceback
+    #traceback.print_stack()
+    print 'stack print, ready for timeout'	
     raise TimeoutError()
-  finally:
-    process.join()
+  return queue.get()
 
 def _runWithTimeout(queue, func, args, kw):
-  import signal
-  import traceback
+   import signal
+   import traceback
+   if hasattr(signal, 'SIGUSR1'):
+     signal.signal(signal.SIGUSR1, lambda sig, stack: traceback.print_stack(stack))
+     print "To see current traceback:"
+     print "  kill -SIGUSR1 %s" % os.getpid()         
+   signal.signal(signal.SIGTERM, lambda sig, stack: traceback.print_stack(stack))
+   queue.put(func(*args, **kw))
 
-  if hasattr(signal, 'SIGUSR1'):
-    signal.signal(signal.SIGUSR1, lambda sig, stack: traceback.print_stack(stack))
-    print "To see current traceback:"
-    print "  kill -SIGUSR1 %s" % os.getpid()
-
-  # print a traceback when terminated.
-  def handler(sig, stack):
-    app.logger.error("Terminating")
-    app.logger.info("".join(traceback.format_stack(stack)))
-    sys.exit(0)
-  signal.signal(signal.SIGTERM, handler)
-
-  queue.put(func(*args, **kw))
 
 @app.route("/runSimulation", methods=["POST", "OPTIONS"])
 def runSimulation():
@@ -154,8 +151,14 @@ def _runSimulation(parameter_dict):
     app.logger.error(tb)
     return dict(error=tb)
 
-def getGUIInstance():
+def getGUIInstance(*args):
     # XXX do not instanciate each time!
+    parser = argparse.ArgumentParser(description='Launch the DREAM simulation platform.')
+    parser.add_argument('gui_class', metavar='GUI_KLASS', nargs="?", default="Default",
+                   help='The GUI klass to launch')
+    arguments = parser.parse_args()
+    global klass_name
+    klass_name = 'dream.simulation.GUI.%s' % arguments.gui_class
     klass = __import__(klass_name, globals(), {}, klass_name)
     instance = klass.Simulation(logger=app.logger)
     return instance
@@ -209,16 +212,14 @@ def main(*args):
                    help='Port number to listen to')
   parser.add_argument('--host', dest='host', default="localhost",
                    help='Host address')
-  parser.add_argument('--logfile', dest='logfile', help='Log to file')
   arguments = parser.parse_args()
   global klass_name
   klass_name = 'dream.simulation.GUI.%s' % arguments.gui_class
-  if arguments.logfile:
-    file_handler = logging.FileHandler(arguments.logfile)
-    file_handler.setLevel(logging.DEBUG)
-    app.logger.addHandler(file_handler)
-
   # start the server
+  file_handler = logging.FileHandler(
+    os.path.join(os.path.dirname(__file__), '..', '..', 'log', 'dream.log'))
+  file_handler.setLevel(logging.DEBUG)
+  app.logger.addHandler(file_handler)
   app.run(debug=True, host=arguments.host, port=arguments.port)
 
 def run(*args):
